complete reconfig
