+++ .gitignore
@@ -1,0 +1,7 @@
+node_modules
+public/bundle.js
+public/bundle.js.map
+secrets.js
+.DS_Store
+npm-debug.log
+yarn-error.log
