+++ client/components/Main.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import TopicsListWithData from './TopicsListWithData';
+// import PropTypes from 'prop-types';
+
+const Main = () => {
+	return (
+		<div>
+			<TopicsListWithData />
+		</div>
+	)
+}
+
+export default Main;
