+++ client/components/index.js
@@ -1,0 +1,10 @@
+/**
+ * `components/index.js` exists simply as a 'central export' for our components.
+ * This way, we can import all of our components from the same place, rather than
+ * having to figure out which file they belong to!
+ */
+export {default as Main} from './Main';
+export {default as TopicDetails} from './TopicDetails';
+export {default as NotFound} from './NotFound';
+/// export {default as UserHome} from './user-home'
+// export {Login, Signup} from './auth-form'
