+++ client/history.js
@@ -1,0 +1,6 @@
+import createHistory from 'history/createBrowserHistory'
+import createMemoryHistory from 'history/createMemoryHistory'
+
+const history = process.env.NODE_ENV === 'test' ? createMemoryHistory() : createHistory()
+
+export default history
