+++ client/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Routes from './routes';
+import './index.scss';
+
+ReactDOM.render(
+	<Routes />,
+  document.getElementById('app')
+);
