--- client/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "client",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "apollo-client": "^1.1.1",
-    "apollo-test-utils": "^0.1.1",
-    "false": "0.0.4",
-    "graphql-tools": "^0.10.0",
-    "react": "^15.4.2",
-    "react-apollo": "^1.1.2",
-    "react-dom": "^15.4.2",
-    "react-router-dom": "^4.0.0"
-  },
-  "devDependencies": {
-    "react-scripts": "0.9.3"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
