--- client/src/App.css
@@ -1,182 +1,0 @@
-html, body {
-  padding: 0;
-  margin: 0;
-  height: 100%;
-  background-image: linear-gradient(175deg, #2b3658 0%, #523e5b 100%);
-}
-
-.App-logo {
-  height: 80px;
-}
-
-.App-header {
-  background-color: #222;
-  height: 150px;
-  padding: 20px;
-  color: white;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-a {
-  text-decoration: none;
-  color: inherit;
-}
-
-div.NotFound {
-  position: relative;
-  padding: 10px 50px;
-  color: rgba(255, 255, 255, 0.8);
-  font-weight: 300;
-  width: 200px;
-  text-align: left;
-  font-size: 16px;
-}
-
-.navbar {
-  background-color: #333;
-  color: #fff;
-  padding: 10px 30px;
-  text-align: left;
-  font-weight: 200;
-  font-size: 20px;
-  line-height: 30px;
-  display: flex;
-  align-items: center;
-}
-
-.navbar:before {
-  content: "";
-  height: 30px;
-  width: 30px;
-  display: inline-block;
-  background-size: contain;
-  margin-right: 10px;
-}
-
-div.channelsList {
-  padding: 0 30px;
-  position: fixed;
-}
-
-div.channel {
-  position: relative;
-  padding: 10px 20px;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  color: rgba(255, 255, 255, 0.8);
-  font-weight: 300;
-  width: 200px;
-  text-align: left;
-  padding-left: 30px;
-  font-size: 16px;
-}
-
-div.optimistic {
-  color: rgba(255, 255, 255, 0.5);
-}
-
-div.channelsList input {
-  width: 200px;
-  background-color: transparent;
-  padding: 10px 20px 10px 30px;
-  outline: none;
-  border: none;
-  color: #fff;
-  font-size: 16px;
-  font-weight: 300;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  transition: 300ms all ease-out;
-}
-
-div.channelsList:before {
-  content: "+";
-  display: inline-block;
-  height: 10px;
-  width: 10px;
-  /*background-color: #fff;*/
-  position: absolute;
-  top: 11px;
-  left: 40px;
-  color: rgba(255, 255, 255, 0.6);
-}
-
-div.channelsList input:focus {
-  border-bottom: 1px solid rgba(255, 255, 255, 0.5);
-  transition: 300ms all ease-out;
-}
-
-div.channel span {
-  position: absolute;
-  left: 10px;
-}
-
-div.channelName {
-  width: 200px;
-  background-color: transparent;
-  padding: 10px 50px 10px 30px;
-  outline: none;
-  border: none;
-  color: #fff;
-  font-size: 32px;
-  text-align: center;
-  color: #22A699;
-}
-
-div.messagesList {
-  padding: 0 30px;
-  position: fixed;
-}
-
-div.message {
-  position: relative;
-  padding: 10px 20px;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  color: rgba(255, 255, 255, 0.8);
-  font-weight: 300;
-  width: 200px;
-  text-align: left;
-  padding-left: 30px;
-  font-size: 16px;
-}
-
-div.optimistic {
-  color: rgba(255, 255, 255, 0.5);
-}
-
-div.messagesList input {
-  width: 200px;
-  background-color: transparent;
-  padding: 10px 20px 10px 30px;
-  outline: none;
-  border: none;
-  color: #fff;
-  font-size: 16px;
-  font-weight: 300;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  transition: 300ms all ease-out;
-}
-
-div.messageInput:before {
-  content: "+";
-  display: inline-block;
-  height: 0px;
-  width: 0px;
-  /*background-color: #fff;*/
-  position: relative;
-  top: -1px;
-  left: 8px;
-  color: rgba(255, 255, 255, 0.6);
-}
-
-div.messagesList input:focus {
-  border-bottom: 1px solid rgba(255, 255, 255, 0.5);
-  transition: 300ms all ease-out;
-}
-
-div.message span {
-  position: absolute;
-  left: 10px;
-}
-
