--- client/src/schema.js
@@ -1,11 +1,0 @@
-export const typeDefs = `
-
-type Category {
-  id: ID!
-  name: String
-}
-
-type Query {
-  categories: [Category]
-}
-`;
