+++ package.json
@@ -1,0 +1,71 @@
+{
+  "name": "definitive-list",
+  "version": "1.0.0",
+  "description": "A list for keeping track of books",
+  "main": "index.js",
+  "scripts": {
+    "build-client": "webpack",
+    "build-client-watch": "webpack -w",
+    "deploy": "script/deploy",
+    "seed": "node script/seed.js",
+    "start": "node server",
+    "start-dev": "npm run build-client-watch & npm run start-server",
+    "start-server": "NODE_ENV='development' nodemon server -e html,js,scss --ignore public --exec babel-node",
+    "test": "NODE_ENV='test' DATABASE_URL='postgres://hocalhost:5432/definitive-list-test' mocha ./**/*.spec.js --compilers js:babel-register"
+  },
+  "author": "elsa",
+  "license": "ISC",
+  "dependencies": {
+    "apollo-cache-inmemory": "^1.1.5",
+    "apollo-client": "^2.2.0",
+    "apollo-client-preset": "^1.0.6",
+    "apollo-engine": "^0.8.3",
+    "apollo-link-http": "^1.3.2",
+    "apollo-server-express": "^1.3.2",
+    "body-parser": "^1.17.1",
+    "connect-session-sequelize": "^5.1.0",
+    "cors": "^2.8.3",
+    "express": "^4.15.2",
+    "express-session": "^1.15.6",
+    "faker": "^4.1.0",
+    "graphql": "^0.9.6",
+    "graphql-server-express": "^0.6.0",
+    "graphql-tag": "^2.6.1",
+    "graphql-tools": "^0.10.1",
+    "history": "^4.7.2",
+    "lodash": "^4.17.4",
+    "morgan": "^1.9.0",
+    "pg": "^6.4.2",
+    "pg-hstore": "^2.3.2",
+    "react": "^16.2.0",
+    "react-apollo": "^2.0.4",
+    "react-dom": "^16.2.0",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
+    "sequelize": "^4.31.1",
+    "sqlite3": "^3.1.13"
+  },
+  "devDependencies": {
+    "babel-cli": "^6.24.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-es2015": "^6.24.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.22.0",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "css-loader": "^0.28.8",
+    "file-loader": "^1.1.6",
+    "mocha": "^4.1.0",
+    "node-sass": "^4.7.2",
+    "nodemon": "^1.11.0",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.19.1",
+    "webpack": "^3.10.0"
+  },
+  "babel": {
+    "presets": [
+      "env"
+    ]
+  }
+}
