+++ public/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link rel="shortcut icon" href="favicon.ico">
+    <title>Booklist</title>
+  </head>
+  <body>
+    <div id="app"></div>
+    <script defer src="bundle.js"></script>
+  </body>
+</html>
