--- resources/App.css
@@ -1,30 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 80px;
-}
-
-.App-header {
-  background-color: #222;
-  height: 150px;
-  padding: 20px;
-  color: white;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-.Item-list {
-  list-style: none;
-}
-
-.Item-list li {
-  border: 1px dotted gray;
-  width: 20%;
-  height: 3em;  
-  margin: auto;
-  text-align: left;
-}
