--- server/.babelrc
@@ -1,7 +1,0 @@
-{
-  "passPerPreset": true,
-  "presets": [
-    "es2015",
-    "stage-0"
-  ]
-}
