+++ server/db/db.js
@@ -1,0 +1,6 @@
+const Sequelize = require('sequelize');
+const db = new Sequelize(
+	process.env.DATABASE_URL || 'postgres://localhost:5432/definitive-list', { logging: false }
+);
+
+module.exports = db;
