+++ server/db/index.js
@@ -1,0 +1,6 @@
+const db = require('./db');
+
+// register models
+require('./models');
+
+module.exports = db;
