+++ server/db/models/index.js
@@ -1,0 +1,6 @@
+const Topic = require('./topic');
+const Item = require('./item');
+
+Item.belongsTo(Topic);
+
+module.exports = { Topic, Item };
