+++ server/db/models/item.js
@@ -1,0 +1,10 @@
+const Sequelize = require('sequelize');
+const db = require('../db');
+
+const Item = db.define('item', {
+	name: {
+		type: Sequelize.STRING
+	}
+});
+
+module.exports = Item;
