+++ server/db/models/topic.js
@@ -1,0 +1,11 @@
+const Sequelize = require('sequelize');
+const db = require('../db');
+
+
+const Topic = db.define('topic', {
+	name: { 
+		type: Sequelize.STRING 
+	}
+});
+
+module.exports = Topic;
