--- server/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "tutorial-server",
-  "version": "1.0.0",
-  "description": "A simple GraphQL server",
-  "main": "server.js",
-  "scripts": {
-    "start": "nodemon ./server.js --exec babel-node -e js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "babel-cli": "^6.24.0",
-    "babel-preset-es2015": "^6.24.0",
-    "babel-preset-stage-0": "^6.22.0",
-    "nodemon": "^1.11.0"
-  },
-  "dependencies": {
-    "body-parser": "^1.17.1",
-    "cors": "^2.8.3",
-    "express": "^4.15.2",
-    "faker": "^4.1.0",
-    "graphql": "^0.9.1",
-    "graphql-server-express": "^0.6.0",
-    "graphql-tools": "^0.10.1",
-    "lodash": "^4.17.4",
-    "sequelize": "^4.31.1",
-    "sqlite3": "^3.1.13"
-  }
-}
